requirements.txt
Flask
Flask-Cors
Flask-JWT-Extended
Flask-SQLAlchemy
fastapi
uvicorn
python-dotenv
requests
pydantic
apscheduler
python-multipart
pydantic-settings
sqlalchemy
slowapi
prometheus-fastapi-instrumentator
sentry-sdk
aiosqlite
circuitbreaker
pandas
phonenumbers